feat: status-page iframe
