+++ apps/docs/src/content/docs/guides/how-to-embed-status-page-iframe.mdx
@@ -1,0 +1,83 @@
+---
+title: How to Embed a Status Page in an Iframe
+description: Embed a trimmed, chromeless version of your openstatus status page on any website using the `?iframe` query parameter.
+---
+
+## Problem
+
+You want to embed your openstatus status page in your marketing site, dashboard, or docs — without the full header, footer, and navigation that the standalone page includes.
+
+## Solution
+
+Append `?iframe` to your status page URL to render a chromeless version suitable for embedding. Use the value to choose exactly which sections to show.
+
+## URL contract
+
+```
+https://<your-slug>.openstatus.dev/?iframe=<sections>
+```
+
+- The **presence** of `?iframe` turns on embed mode. The header, footer, and subscribe form are hidden.
+- The **value** is a comma-separated list of sections to show.
+- An **empty value** (`?iframe` or `?iframe=`) shows all sections.
+
+### Sections
+
+| Key | Renders |
+|---|---|
+| `title` | Page title + description |
+| `banner` | Current incident / maintenance banner |
+| `components` | Component health trackers |
+| `feed` | Past status reports and maintenances |
+
+Unknown keys are silently ignored. Section names are case-insensitive. If nothing valid is provided (e.g. `?iframe=,` or `?iframe=unknown`), all sections are shown.
+
+Section keys apply to the **root status page URL only**. Other routes (`/monitors`, `/events`, `/events/report/:id`) accept `?iframe` to hide header/footer chrome, but their content is unchanged — they render normally inside the frame.
+
+### Additional parameters
+
+| Param | Values | Behavior |
+|---|---|---|
+| `theme` | `light`, `dark` | Forces the theme. Honored only when `?iframe` is present. Falls back to system preference, then the page's configured default. |
+
+## Examples
+
+Show the full chromeless page:
+
+```html
+<iframe
+  src="https://status.example.com/?iframe"
+  width="100%"
+  height="600"
+  style="border: 0;"
+></iframe>
+```
+
+Show only the title and current banner:
+
+```html
+<iframe
+  src="https://status.example.com/?iframe=title,banner"
+  width="100%"
+  height="300"
+  style="border: 0;"
+></iframe>
+```
+
+Show components in forced dark theme:
+
+```html
+<iframe
+  src="https://status.example.com/?iframe=components&theme=dark"
+  width="100%"
+  height="400"
+  style="border: 0;"
+></iframe>
+```
+
+## Notes and caveats
+
+- **Height:** iframes do not auto-size to their content. Set a `height` attribute appropriate for the sections you enabled.
+- **Internal links** (e.g. clicking an incident banner) open in a new tab when embedded, so your site's iframe stays put.
+- **Password-protected status pages** may not authenticate correctly inside iframes due to browser third-party cookie restrictions (Safari ITP, Chrome). Visitors may need to authenticate on the full status page first.
+- **Sandboxed iframes** must include `allow-popups` for the "open in new tab" behavior to work.
