+++ apps/status-page/src/app/(status-page)/[domain]/[locale]/(public)/search-params.ts
@@ -1,0 +1,6 @@
+import { iframeParser } from "@/lib/iframe-params";
+import { createSearchParamsCache } from "nuqs/server";
+
+export const iframeSearchParamsCache = createSearchParamsCache({
+  iframe: iframeParser,
+});
