+++ apps/status-page/src/app/not-found.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import Link from "next/link";
 import { useRouter } from "next/navigation";
 
+import { Link } from "@/components/common/link";
 import { ThemeProvider } from "@/components/themes/theme-provider";
 import { Button } from "@openstatus/ui/components/ui/button";
 
@@ -34,7 +34,9 @@ export default function NotFound() {
                   Go Back
                 </Button>
                 <Button size="lg" asChild>
-                  <Link href="/">Home</Link>
+                  <Link variant="unstyled" href="/">
+                    Home
+                  </Link>
                 </Button>
               </div>
             </div>
